+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ﻿## Version 4.09 [In Development]
 
+**All WoW Versions:**  
+- Added back the Augmentation/Support layout (Centered Raidframes, No Actionbars)  
+
 **Retail:**  
 - Fixed Private Aura positions (Raid1, Raid2, Raid3)  
 - Updated BetterCooldownManager (Fixes the initial position for 1080p users)